Renamed MediapartProxy to SiteReaderProxy
