--- src/BD/MediapartProxyBundle/BDMediapartProxyBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace BD\MediapartProxyBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class BDMediapartProxyBundle extends Bundle
-{
-}
