+++ src/BD/SiteReaderProxyBundle/BDSiteReaderProxyBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace BD\SiteReaderProxyBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class BDSiteReaderProxyBundle extends Bundle
+{
+}
