+++ src/BD/SiteReaderProxyBundle/DependencyInjection/BDSiteReaderProxyExtension.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BD\MediapartProxyBundle\DependencyInjection;
+namespace BD\SiteReaderProxyBundle\DependencyInjection;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
@@ -12,7 +12,7 @@ use Symfony\Component\DependencyInjection\Loader;
  *
  * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
  */
-class BDMediapartProxyExtension extends Extension
+class BDSiteReaderProxyExtension extends Extension
 {
     /**
      * {@inheritDoc}
