+++ src/BD/SiteReaderProxyBundle/Resources/config/sites_routing/lmd.yml
@@ -1,0 +1,5 @@
+bd_site_reader_proxy.site.lmd.all:
+    path: /{uri}
+    defaults: { _controller: bd_site_reader_proxy.site.controller.lmd:uriAction }
+    requirements:
+        uri: .+
