+++ src/BD/SiteReaderProxyBundle/Resources/config/sites_routing/mediapart.yml
@@ -1,0 +1,9 @@
+bd_site_reader_proxy.mediapart_articles_feed:
+    path: /articles/feed
+    defaults: { _controller: bd_site_reader_proxy.controller.feed_proxy:proxyArticlesFeedAction }
+
+bd_site_reader_proxy.mediapart_article:
+    path: /journal/{uri}
+    defaults: { _controller: bd_site_reader_proxy.controller.feed_proxy:proxyArticleAction }
+    requirements:
+        uri: .+
