+++ src/BD/SiteReaderProxyBundle/Resources/translations/messages.fr.xlf
